improved testing
