+++ tests/resources/file_caching/test_cache_metrics_by_city.py
